requirements.txt
Django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
drf-spectacular==0.28.0
psycopg2-binary==2.9.9
django-cors-headers==4.6.0
Pillow==11.1.0
python-decouple==3.8
uvicorn[standard]
channels==4.2.0
channels-redis==4.2.0